move all duplicated modules in shared moduls
