+++ src/app/components/mainpage/mainpage.module.ts
@@ -6,27 +6,9 @@ import { TodoListPageComponent } from './todo-list-page/todo-list-page.component
 import { WishListPageComponent } from './wish-list-page/wish-list-page.component';
 import { ProfilePageComponent } from './profile-page/profile-page.component';
 
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { MatTableModule } from '@angular/material/table';
-import { MatPaginatorModule } from '@angular/material/paginator';
-import { MatSortModule } from '@angular/material/sort';
-import { MatCheckboxModule } from '@angular/material/checkbox';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { MatIconModule } from '@angular/material/icon';
-import { MatButtonModule } from '@angular/material/button';
-import { MatDialogModule } from '@angular/material/dialog';
-import { MatFormFieldModule } from '@angular/material/form-field';
-import { MatInputModule } from '@angular/material/input';
-import { MatSelectModule } from '@angular/material/select';
-import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { FormsModule } from '@angular/forms';
+import { SharedModule } from 'src/app/shared/shared.module';
 
-import { ImageCropperModule } from 'ngx-image-cropper';
-import { LottieModule } from 'ngx-lottie';
-import player from 'lottie-web';
-
-export function playerFactory() {
-  return player;
-}
 
 @NgModule({
   declarations: [
@@ -37,22 +19,8 @@ export function playerFactory() {
   imports: [
     CommonModule,
     FormsModule,
-    ReactiveFormsModule,
     MainpageRoutingModule,
-    MatButtonModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatSelectModule,
-    MatIconModule,
-    MatTableModule,
-    MatPaginatorModule,
-    MatSortModule,
-    MatCheckboxModule,
-    MatDialogModule,
-    MatProgressBarModule,
-    MatSnackBarModule,
-    ImageCropperModule,
-    [LottieModule.forRoot({ player: playerFactory })]
+    SharedModule
   ]
 })
 export class MainpageModule { }
